Fix the "reset filters" button
In previous releases of ember-submission-form-fields each field had both an "edit" and "show" version. We could (mis)use this behaviour to reset the form fields by temporarily switching the "@show" argument value.

Since a different component was being rendered in the read only mode, Ember throws away the edit component instances and recreates them when the "@show" flag was turned off. This had the effect that all form fields would reset.

When we deduplicated all of the form fields this trick no longer works since Ember doesn't destroy the component instances.

As a quick workaround we can temporarily render a new `RdfForm` instance so that all the components will be destroyed and recreated.